refactor: implement folder structure
