+++ src/app.js
@@ -1,0 +1,14 @@
+const express = require('express');
+const cors = require('cors');
+const bodyParser = require('body-parser');
+const uploadRoutes = require('./routes/uploadRoutes');
+
+const app = express();
+
+app.use(cors());
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: true }));
+
+app.use('/', uploadRoutes);
+
+module.exports = app;
